Working on refactoring options parsing
