+++ cmd/soroban-rpc/internal/daemon/daemon.go
@@ -96,7 +96,7 @@ func (d *Daemon) Close() error {
 }
 
 // newCaptiveCore creates a new captive core backend instance and returns it.
-func newCaptiveCore(cfg *config.LocalConfig, logger *supportlog.Entry) (*ledgerbackend.CaptiveStellarCore, error) {
+func newCaptiveCore(cfg *config.Config, logger *supportlog.Entry) (*ledgerbackend.CaptiveStellarCore, error) {
 	httpPortUint := uint(cfg.CaptiveCoreHTTPPort)
 	var captiveCoreTomlParams ledgerbackend.CaptiveCoreTomlParams
 	captiveCoreTomlParams.HTTPPort = &httpPortUint
@@ -124,7 +124,7 @@ func newCaptiveCore(cfg *config.LocalConfig, logger *supportlog.Entry) (*ledgerb
 
 }
 
-func MustNew(cfg config.LocalConfig, endpoint string, adminEndpoint string) *Daemon {
+func MustNew(cfg config.Config) *Daemon {
 	logger := supportlog.New()
 	logger.SetLevel(cfg.LogLevel)
 
@@ -223,7 +223,7 @@ func MustNew(cfg config.LocalConfig, endpoint string, adminEndpoint string) *Dae
 	preflightWorkerPool := preflight.NewPreflightWorkerPool(
 		cfg.PreflightWorkerCount, cfg.PreflightWorkerQueueSize, ledgerEntryReader, cfg.NetworkPassphrase, logger)
 
-	jsonRPCHandler := internal.NewJSONRPCHandler(&cfg, internal.HandlerParams{
+	jsonRPCHandler := internal.NewJSONRPCHandler(&cfg.DaemonConfig, internal.HandlerParams{
 		EventStore:       eventStore,
 		TransactionStore: transactionStore,
 		Logger:           logger,
@@ -244,11 +244,11 @@ func MustNew(cfg config.LocalConfig, endpoint string, adminEndpoint string) *Dae
 	daemon.jsonRPCHandler = &jsonRPCHandler
 
 	daemon.server = &http.Server{
-		Addr:        endpoint,
+		Addr:        cfg.Endpoint,
 		Handler:     httpHandler,
 		ReadTimeout: defaultReadTimeout,
 	}
-	if adminEndpoint != "" {
+	if cfg.AdminEndpoint != "" {
 		adminMux := supporthttp.NewMux(logger)
 		adminMux.HandleFunc("/debug/pprof/", pprof.Index)
 		adminMux.HandleFunc("/debug/pprof/cmdline", pprof.Cmdline)
@@ -256,7 +256,7 @@ func MustNew(cfg config.LocalConfig, endpoint string, adminEndpoint string) *Dae
 		adminMux.HandleFunc("/debug/pprof/symbol", pprof.Symbol)
 		adminMux.HandleFunc("/debug/pprof/trace", pprof.Trace)
 		adminMux.Handle("/metrics", promhttp.HandlerFor(metricsRegistry, promhttp.HandlerOpts{}))
-		daemon.adminServer = &http.Server{Addr: adminEndpoint, Handler: adminMux}
+		daemon.adminServer = &http.Server{Addr: cfg.AdminEndpoint, Handler: adminMux}
 	}
 	daemon.registerMetrics()
 	return daemon
